+++ internal/catalog/cl_teku_test.go
@@ -1,8 +1,0 @@
-package catalog
-
-import "testing"
-
-func TestTeku(t *testing.T) {
-	tr := newTestingFramework("teku")
-	tr.ImageExists(t)
-}