+++ internal/catalog/el_besu_test.go
@@ -1,8 +1,0 @@
-package catalog
-
-import "testing"
-
-func TestBesu(t *testing.T) {
-	tr := newTestingFramework("besu")
-	tr.ImageExists(t)
-}