+++ internal/catalog/el_geth_test.go
@@ -1,8 +1,0 @@
-package catalog
-
-import "testing"
-
-func TestGeth(t *testing.T) {
-	tr := newTestingFramework("geth")
-	tr.ImageExists(t)
-}