+++ internal/client/runner/driver/driver.go
@@ -13,7 +13,7 @@ type Driver interface {
 	WaitTask(ctx context.Context, taskID string) (<-chan *proto.ExitResult, error)
 	StopTask(taskID string, timeout time.Duration) error
 	DestroyTask(taskID string, force bool) error
-	CreateNetwork(allocID string, hostname string) (*proto.NetworkSpec, bool, error)
+	CreateNetwork(allocID string, dnsAlias []string, hostname string) (*proto.NetworkSpec, bool, error)
 	DestroyNetwork(spec *proto.NetworkSpec) error
 	ExecTask(taskID string, cmd []string) (*ExecTaskResult, error)
 }