+++ internal/cmd/deploy.go
@@ -23,6 +23,10 @@ type DeployCommand struct {
 	params string
 
 	metrics bool
+
+	alias string
+
+	logLevel string
 }
 
 // Help implements the cli.Command interface
@@ -45,6 +49,8 @@ func (c *DeployCommand) Run(args []string) int {
 	flags.StringVar(&c.allocId, "alloc", "", "")
 	flags.StringVar(&c.params, "params", "", "")
 	flags.BoolVar(&c.metrics, "metrics", true, "")
+	flags.StringVar(&c.alias, "alias", "", "")
+	flags.StringVar(&c.logLevel, "log-level", "info", "")
 
 	if err := flags.Parse(args); err != nil {
 		c.UI.Error(err.Error())
@@ -92,6 +98,8 @@ func (c *DeployCommand) Run(args []string) int {
 		AllocationId: c.allocId,
 		Chain:        c.chain,
 		Metrics:      c.metrics,
+		Alias:        c.alias,
+		LogLevel:     c.logLevel,
 	}
 	resp, err := clt.Apply(context.Background(), req)
 	if err != nil {