+++ internal/server/server.go
@@ -102,24 +102,33 @@ func (s *Server) Stop() {
 	s.grpcServer.Stop()
 }
 
-func (s *Server) Create(req *proto.ApplyRequest, input map[string]interface{}) (string, error) {
-	deployableTasks, err := s.catalog.Build(req, input)
+func (s *Server) Create(req *proto.ApplyRequest) (string, error) {
+	allocId := req.AllocationId
+
+	var alloc *proto.Allocation
+	var prevState []byte
+
+	if allocId != "" {
+		// load allocation from the state
+		var err error
+
+		if alloc, err = s.state.GetAllocation(allocId); err != nil {
+			return "", err
+		}
+		prevState = alloc.InputState
+	}
+
+	state, deployableTasks, err := s.catalog.Build(prevState, req)
 	if err != nil {
 		return "", fmt.Errorf("failed to run plugin '%s': %v", req.Action, err)
 	}
 
-	allocId := req.AllocationId
-
-	if allocId != "" {
+	if alloc != nil {
 		// update the deployment
-		alloc, err := s.state.GetAllocation(allocId)
-		if err != nil {
-			return "", err
-		}
-
 		alloc = alloc.Copy()
 		alloc.Sequence++
 		alloc.Tasks = deployableTasks
+		alloc.InputState = state
 
 		if err := s.state.UpsertAllocation(alloc); err != nil {
 			return "", err
@@ -128,9 +137,11 @@ func (s *Server) Create(req *proto.ApplyRequest, input map[string]interface{}) (
 		allocId = uuid.Generate()
 
 		alloc := &proto.Allocation{
-			Id:     allocId,
-			NodeId: "local",
-			Tasks:  deployableTasks,
+			Id:         allocId,
+			NodeId:     "local",
+			Tasks:      deployableTasks,
+			InputState: state,
+			Alias:      req.Alias,
 		}
 		if err := s.state.UpsertAllocation(alloc); err != nil {
 			return "", err