+++ internal/server/server_test.go
@@ -11,12 +11,14 @@ import (
 )
 
 type dummyCatalog struct {
-	tasks map[string]*proto.Task
+	prev       []byte
+	createTask *proto.Task
 }
 
-func (d *dummyCatalog) Build(req *proto.ApplyRequest, input map[string]interface{}) (map[string]*proto.Task, error) {
+func (d *dummyCatalog) Build(prev []byte, req *proto.ApplyRequest) ([]byte, map[string]*proto.Task, error) {
 	// this is enough to generate an allocation
-	return d.tasks, nil
+	d.prev = prev
+	return req.Input, map[string]*proto.Task{"task": d.createTask}, nil
 }
 
 func TestCreate(t *testing.T) {
@@ -32,33 +34,42 @@ func TestCreate(t *testing.T) {
 	srv, _ := NewServer(hclog.NewNullLogger(), cfg)
 	srv.catalog = catalog
 
-	catalog.tasks = map[string]*proto.Task{}
-	catalog.tasks["a"] = &proto.Task{
+	catalog.createTask = &proto.Task{
 		Args: []string{"a"},
 	}
 
-	allocid, err := srv.Create(&proto.ApplyRequest{}, nil)
+	input := []byte{0x1, 0x2, 0x3}
+
+	allocid, err := srv.Create(&proto.ApplyRequest{Input: input})
 	require.NoError(t, err)
+
+	// 'prev' is empty since there was no previous state
+	require.Empty(t, catalog.prev)
 
 	// the allocation should be on the state with sequence=0
 	alloc, err := srv.state.GetAllocation(allocid)
 	require.NoError(t, err)
 	require.Equal(t, alloc.Id, allocid)
 	require.Equal(t, alloc.Sequence, int64(0))
-	require.Len(t, alloc.Tasks, 1)
+	require.Equal(t, alloc.Tasks["task"].Args, []string{"a"})
+	require.Equal(t, alloc.InputState, input)
 
 	// update the allocation
-	catalog.tasks["b"] = &proto.Task{
+	catalog.createTask = &proto.Task{
 		Args: []string{"b"},
 	}
 
-	allocid2, err := srv.Create(&proto.ApplyRequest{AllocationId: allocid}, nil)
+	input2 := []byte{0x4, 0x5, 0x6}
+
+	allocid2, err := srv.Create(&proto.ApplyRequest{Input: input2, AllocationId: allocid})
 	require.NoError(t, err)
 	require.Equal(t, allocid, allocid2)
+	require.Equal(t, catalog.prev, input)
 
 	// the allocation shoulld be on the state with sequence=1
 	alloc, err = srv.state.GetAllocation(allocid)
 	require.NoError(t, err)
 	require.Equal(t, alloc.Sequence, int64(1))
-	require.Len(t, alloc.Tasks, 2)
+	require.Equal(t, alloc.Tasks["task"].Args, []string{"b"})
+	require.Equal(t, alloc.InputState, input2)
 }