+++ CHANGELOG.md
@@ -1,5 +1,13 @@
-# 0.1.1 (Unreleased)
+# 0.1.2 (Unreleased)
 
+IMPROVEMENTS:
+
+- catalog: Add logging field in catalog [[GH-62](https://github.com/umbracle/eth2-validator/issues/62)]
+- core: Add an alias for the deployment [[GH-23](https://github.com/umbracle/eth2-validator/issues/23)]
+
+# 0.1.1 (18 April, 2023)
+
+- feat: Use Id or Prefix in `deployment status` [[GH-22](https://github.com/umbracle/eth2-validator/issues/22)]
 - feat: List available chains in the plugin [[GH-48](https://github.com/umbracle/eth2-validator/issues/48)]
 - feat: Add state persistence [[GH-8](https://github.com/umbracle/eth2-validator/issues/8)]
 - feat: Add cli `catalog list` and `catalog inspect` commands [[GH-6](https://github.com/umbracle/eth2-validator/issues/6)]